requirements.txt
Django
djangorestframework
django-registration
BeautifulSoup
mysqlclient==1.3.4
virtualenv
requests
numpy
pandas
python-dateutil
redis
PyPDF2
django-bootstrap-ui